+++ UserManager/src/main/java/picto/com/usermanager/domain/user/dto/request/SignUpRequest.java
@@ -13,14 +13,14 @@ public class SignUpRequest {
     private String password;
     private String name;
     // 유저값을 클라이언트에서 부여
-    private int userId;
+    private Long userId;
 
     private double lat;
     private double lng;
 
 
     @Builder
-    public SignUpRequest(String email, String password, String name, int userId, double lat, double lng) {
+    public SignUpRequest(String email, String password, String name, Long userId, double lat, double lng) {
         this.email = email;
         this.password = password;
         this.name = name;
